split main file to seperate typeDefs and resolvers files
