+++ part8/library-backend/index.js
@@ -1,0 +1,42 @@
+const { ApolloServer, } = require("apollo-server");
+const mongoose = require("mongoose");
+const jwt = require("jsonwebtoken");
+const typeDefs = require("./typeDefs");
+const resolvers = require("./resolvers");
+
+require("dotenv").config();
+const MONGODB_URI = process.env.MONGODB_URI;
+const JWT_SECRET = process.env.JWT_SECRET;
+
+console.log("Connecting to MongoDB");
+
+mongoose
+  .connect(MONGODB_URI, {
+    useNewUrlParser: true,
+    useUnifiedTopology: true,
+    useFindAndModify: false,
+  })
+  .then(() => {
+    console.log("Connected to MongoDB");
+  })
+  .catch((err) => {
+    console.log("Error connecting to MongoDB", err.message);
+  });
+
+const server = new ApolloServer({
+  typeDefs,
+  resolvers,
+  context: async ({ req }) => {
+    const auth = req ? req.headers.authorization : null;
+    if (auth && auth.toLowerCase().startsWith("bearer")) {
+      const decoded = jwt.verify(auth.substring(7), JWT_SECRET);
+      const currentUser = await User.findById(decoded.id);
+      return { currentUser };
+    }
+  },
+});
+
+server.listen().then(({ url, subscriptionsUrl }) => {
+  console.log(`Server ready at ${url}`);
+  console.log(`Subscriptions ready at ${subscriptionsUrl}`);
+});
